firebase_demo_brief/requirements.txt
Flask
openai
python-dotenv
airtable-python-wrapper
